Added note about Android for Maven Eclipse.
The Android for Maven Eclipse (m2e-android) plugin now includes support for
Robolectric and other unit testing frameworks. Additionally, it now allows you
to have unit tests and application code in the same project:

https://github.com/rgladwell/m2e-android/issues/52

This commit adds a note for Eclipse users to this effect, pointing them in the
direction of the plugin.