Update the "Building Robolectric" page
This PR updates the [Building Robolectric](https://robolectric.org/building-robolectric/) page.

I've changed some wordings, and updated the commands to build and test Robolectric.
These come from the Readme file on the main project:
https://github.com/robolectric/robolectric/tree/master?tab=readme-ov-file#building-and-contributing

Once this PR is merged, I suggest updating the Readme on the main repo to replace the
"Building And Contributing" section with a link to the website. This is to avoid having duplicated
information to maintain in different places.